# back/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 문해한 하루 API
  description: 문해한 하루 API 문서입니다.
  version: 0.1.9
servers:
  - url: https://elice-kdt-ai-4th-team02.elicecoding.com
    description: Server(production)
  - url: http://localhost:5001
    description: Server(dev)
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 6273ed6490b074ccd0453dd9
        nickname: 
          type: string
          example: 나는abcde
        email: 
          type: string
          example: "abcde@abc.com"
        badge: 
          type: array
          items: {}
        level:
          type: integer
          example: 1
        point:
          type: integer
          example: 10
        postLikes:
          type: array
          items: {}
        profileUrl:
          type: string
          example: "https://team2.cdn.ntruss.com/users/default.png"
        kakaoId:
          type: integer
          example: 12345678
        introduce:
          type: string
          example: "안녕하세요 나는abcde입니다."
      # Both properties are required
      required:  
        - id
        - nickname
        - email
        - password
    UserRegister:
      type: object
      properties:
        email: 
          type: string
          example: "abcdef@abc.com"
        password:
          type: string
          example: abcd1234
        nickname:
          type: string
          example: 나는abcdef
      required:
        - email
        - password
        - nickname
    UserLogin:
      type: object
      properties:
        email: 
          type: string
          example: "abcdef@abc.com"
        password:
          type: string
          example: abcd1234
      required:
        - email
        - password
...
  parameters:
    sortFieldParams:
      name: sortField
      in: query
      required: false
      schema:
        type: string
      description: 정렬 field
    sortTypeParams:
      name: sortType
      in: query
      required: false
      schema:
        type: string
      description: 정렬 타입
    pageParams:
      name: page
      in: query
      required: false
      schema:
        type: integer
      description: 페이지 번호
    limitParams:  
      name: limit
      in: query
      required: false
      schema:
        type: integer
      description: 한 페이지에 보여줄 데이터 개수
    userParams:
      name: userId
      in: path
      required: true
      description: user의 등록 id입니다. (objectId)
      schema:
        type : string
        example: 6273ed6490b074ccd0453dd9
...
paths:
...
  /users:
    get:
      parameters:
        - $ref: '#/components/parameters/sortFieldParams'
        - $ref: '#/components/parameters/sortTypeParams'
        - $ref: '#/components/parameters/pageParams'
        - $ref: '#/components/parameters/limitParams'
      summary: 전체 user 정보
      description: 전체 user의 정보를 받아옵니다.
      
      responses:
        '200':    # status code
          description: 전체 user 객체를 배열로 받아옵니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 데이터 리딩시 오류.
        default:
          description: Unexpected error
      tags:
        - users
  /users/register:
    post:
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses: 
        '201':
          description: 회원가입 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - users
  /users/login:
    post:
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses: 
        '201':
          description: 회원가입 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - users
  /oauth/kakao:
    get:
      summary: 카카오로그인
      parameters:
        - in: query
          required: true
          name: code
          schema: 
            type: string
          description: 카카오 로그인 API에서 받은 인가 코드
      responses: 
        '201':
          description: 회원가입 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '200':
          description: 로그인 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - users
  /users/current:
    get:
      summary: 현재 로그인 유저 정보
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 로그인 유저 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - users
  /users/{userId}:
    get:
      summary: 해당 user 조회
      parameters:
        - $ref: '#/components/parameters/userParams'
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - users
    put:
      summary: 해당 user 정보 수정
      parameters:
        - $ref: '#/components/parameters/userParams'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  example: 나는abcde
              
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - users
    delete:
      summary: 해당 user 정보 삭제(탈퇴)
      parameters:
        - $ref: '#/components/parameters/userParams'
      security:
        - bearerAuth: []        
      responses: 
        '200':
          description: success
      tags:
        - users
  /users/{userId}/uploadImage:
    post:
      summary: 해당 user의 프로필 이미지 변경
      parameters:
        - $ref: '#/components/parameters/userParams'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                filename:  # Part 3 (an image)
                  type: string
                  format: binary
                  description: 등록할 이미지(파일)
                prevImage:
                  type: string
                  description: 이전 이미지 url(존재시)
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: string
                    example: success
      tags:
        - users
  /users/{userId}/removeImage:
    patch:
      summary: 해당 user의 이미지 삭제
      parameters:
        - $ref: '#/components/parameters/userParams'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                  example: 1234.png
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: string
                    example: success
      tags:
        - users
  
  # post API      
...
  /userwords:
    post:
      summary: 단어 퀴즈 진행도 저장(마지막 푼 단어 저장)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              example: 사흘
              
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: success
                  word:
                    type: string
                    example: 사흘
      tags:
        - userWord
  /users/{userId}/userword:
    get:
      summary: 해당 유저의 단어 퀴즈 진행도(마지막 푼 단어) 조회
      parameters:
        - $ref: '#/components/parameters/userParams'
      responses: 
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  word:
                    type: string
                    example: 사흘
      tags:
        - userWord
        
  # chart API
...